warning fixes for visit_vtk directory
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@23631 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/visit_vtk/full/vtkOnionPeelFilter.C
@@ -967,7 +967,7 @@ vtkOnionPeelFilter::FindCellsCorrespondingToOriginal(vtkDataSet *input,
         for (int i = comp; i < n; i+=nc )
         {
             int id = i / nc;
-            if (oc[i] == orig && group->IsId(id) == -1)
+            if (oc[i] == (unsigned int)orig && group->IsId(id) == -1)
                 group->InsertNextId(id);
         }
     }
